+++ test/domain/user/use_case/remove_my_page_feed_use_case_test.dart
@@ -41,7 +41,7 @@ void main() {
         final result = await useCase.execute(id: testId);
 
         // Then
-        expect(result, null);
+        expect(result, true);
       });
     });
   });
